feat(schedule)!: supports detecting file paths
